pyproject.toml
[project]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Kunaal",email = "kunaalsai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.2.3,<3.0.0)"
]

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
